fix: setup bash file

+++ others/setup.sh
@@ -1,12 +1,18 @@
 # build docker and install package
-docker-compose build
-docker-compose run --rm server poetry install
-docker-compose run --rm admin-server poetry install
-docker-compose run --rm client npm install
+docker compose build
+docker compose run --rm server poetry install
+docker compose run --rm admin-server poetry install
+docker compose run --rm client npm install
 
-# create .env file from example.env file
-cp server/example.env server/.env
-cp admin_server/example.env server/.env
+# create .env file from example.env file if there are not exist .env file
+if [ ! -f ./server/.env ]; then
+    echo "Generate server .env file"
+    cp server/example.env server/.env
+fi
+if [ ! -f ./admin_server/.env ]; then
+    echo "Generate admin server .env file"
+    cp admin_server/example.env admin_server/.env
+fi
 
 # install common tools
 bun install
